ui test: move settings pages into settings package
