+++ src/test/java/org/synyx/urlaubsverwaltung/ui/ApplicationForLeaveUIIT.java
@@ -23,8 +23,8 @@ import org.synyx.urlaubsverwaltung.ui.pages.ApplicationPage;
 import org.synyx.urlaubsverwaltung.ui.pages.LoginPage;
 import org.synyx.urlaubsverwaltung.ui.pages.NavigationPage;
 import org.synyx.urlaubsverwaltung.ui.pages.OverviewPage;
-import org.synyx.urlaubsverwaltung.ui.pages.SettingsWorkingTimePage;
-import org.synyx.urlaubsverwaltung.ui.pages.SettingsPage;
+import org.synyx.urlaubsverwaltung.ui.pages.settings.SettingsWorkingTimePage;
+import org.synyx.urlaubsverwaltung.ui.pages.settings.SettingsPage;
 import org.synyx.urlaubsverwaltung.workingtime.FederalState;
 import org.synyx.urlaubsverwaltung.workingtime.WorkingTimeWriteService;
 import org.testcontainers.junit.jupiter.Container;
