+++ src/test/java/org/synyx/urlaubsverwaltung/ui/CustomAbsenceTypeUIIT.java
@@ -18,8 +18,8 @@ import org.synyx.urlaubsverwaltung.ui.extension.UiTest;
 import org.synyx.urlaubsverwaltung.ui.pages.ApplicationPage;
 import org.synyx.urlaubsverwaltung.ui.pages.LoginPage;
 import org.synyx.urlaubsverwaltung.ui.pages.NavigationPage;
-import org.synyx.urlaubsverwaltung.ui.pages.SettingsAbsenceTypesPage;
-import org.synyx.urlaubsverwaltung.ui.pages.SettingsPage;
+import org.synyx.urlaubsverwaltung.ui.pages.settings.SettingsAbsenceTypesPage;
+import org.synyx.urlaubsverwaltung.ui.pages.settings.SettingsPage;
 import org.synyx.urlaubsverwaltung.workingtime.WorkingTimeWriteService;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
