+++ src/test/java/org/synyx/urlaubsverwaltung/ui/OvertimeUIIT.java
@@ -18,8 +18,8 @@ import org.synyx.urlaubsverwaltung.ui.pages.LoginPage;
 import org.synyx.urlaubsverwaltung.ui.pages.NavigationPage;
 import org.synyx.urlaubsverwaltung.ui.pages.OvertimeDetailPage;
 import org.synyx.urlaubsverwaltung.ui.pages.OvertimePage;
-import org.synyx.urlaubsverwaltung.ui.pages.SettingsPage;
-import org.synyx.urlaubsverwaltung.ui.pages.SettingsWorkingTimePage;
+import org.synyx.urlaubsverwaltung.ui.pages.settings.SettingsPage;
+import org.synyx.urlaubsverwaltung.ui.pages.settings.SettingsWorkingTimePage;
 import org.synyx.urlaubsverwaltung.workingtime.WorkingTimeWriteService;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
