+++ src/test/java/org/synyx/urlaubsverwaltung/ui/pages/settings/SettingsAbsenceTypesPage.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.ui.pages;
+package org.synyx.urlaubsverwaltung.ui.pages.settings;
 
 import com.microsoft.playwright.Locator;
 import com.microsoft.playwright.Page;
