+++ src/test/java/org/synyx/urlaubsverwaltung/ui/pages/settings/SettingsSubNavigation.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.ui.pages;
+package org.synyx.urlaubsverwaltung.ui.pages.settings;
 
 import com.microsoft.playwright.Page;
 import com.microsoft.playwright.Response;
